Fill db fix. population_to_maintable() to fill_pop command.
