+++ simple_prognozer/mainapp/models.py
@@ -19,6 +19,7 @@ class Subdivision(models.Model):
     # Альтернативное название страны/региона для поиска кол-ва населения
     alias_for_population = models.CharField(verbose_name='Alias',
                                             max_length=128,
+                                            null=True
                                             )
     # FIPS только для US
     fips = models.IntegerField(verbose_name='FIPS US inner number',
