+++ simple_prognozer/mainapp/urls.py
@@ -1,5 +1,3 @@
-
-from django.contrib import admin
 from django.urls import path
 import mainapp.views as mainapp
